Rename wwtranslatedb to WWTranslateDb
