+++ Code/Commando/DlgRestart.cpp
@@ -38,7 +38,7 @@
 #include "Resource.h"
 #include "MainLoop.h"
 #include "String_IDs.h"
-#include <WWTranslateDB\TranslateDB.h>
+#include <WWTranslateDb\TranslateDB.h>
 #include <WWDebug\WWDebug.h>
 
 /******************************************************************************
