+++ Code/Commando/DlgWOLWait.cpp
@@ -39,7 +39,7 @@
 #include "String_ids.h"
 #include <WWOnline\WOLSession.h>
 #include <WWUI\MouseMgr.h>
-#include <WWTranslateDB\TranslateDB.h>
+#include <WWTranslateDb\TranslateDB.h>
 #include <WWDebug\WWDebug.h>
 #include "gamespyadmin.h"
