+++ Code/Commando/WOLChatMgr.cpp
@@ -42,7 +42,7 @@
 #include <WWOnline\WOLChannel.h>
 #include <WWOnline\WOLProduct.h>
 #include "String_IDs.h"
-#include <WWTranslateDB\TranslateDB.h>
+#include <WWTranslateDb\TranslateDB.h>
 #include <WWDebug\WWDebug.h>
 
 using namespace WWOnline;
