+++ Code/Commando/WOLJoinGame.cpp
@@ -45,7 +45,7 @@
 #include <WWOnline\WOLProduct.h>
 #include <WWOnline\WOLChannel.h>
 #include "String_IDs.h"
-#include <WWTranslateDB\TranslateDB.h>
+#include <WWTranslateDb\TranslateDB.h>
 #include <WWDebug\WWDebug.h>
 
 using namespace WWOnline;
