+++ Code/Commando/WOLLoginProfile.cpp
@@ -43,7 +43,7 @@
 #include <WWUI\ListCtrl.h>
 #include <WWLib\Registry.h>
 #include "String_IDs.h"
-#include <WWTranslateDB\TranslateDB.h>
+#include <WWTranslateDb\TranslateDB.h>
 #include <stdio.h>
 
 using namespace WWOnline;
