+++ Code/WWTranslateDb/stringtwiddler.cpp
