+++ Code/WWTranslateDb/stringtwiddler.h
