+++ Code/WWTranslateDb/tdbcategories.h
