+++ Code/WWTranslateDb/tdbcategory.cpp
