+++ Code/WWTranslateDb/tdbcategory.h
