+++ Code/WWTranslateDb/translatedb.cpp
