+++ Code/WWTranslateDb/translatedb.h
