+++ Code/WWTranslateDb/translateobj.cpp
