+++ Code/WWTranslateDb/translateobj.h
