+++ Code/WWTranslateDb/wwtranslatedb.vcxproj
